fix github-actions

+++ make.ps1
@@ -2,21 +2,21 @@
 ##############################################################################################################
 
 Function Show-Usage {
-    Return "
+    "
 Usage: pwsh -File $($PSCommandPath) [OPTIONS]
 Options:
     build   Build program
-"
+" | Out-Host
 }
 
 Function Request-File {
-    ForEach ($REPLY in $args) {
-        $params = @{
-            Uri = $REPLY
-            OutFile = (Split-Path -Path $REPLY -Leaf).Split('?')[0]
+    While ($Input.MoveNext()) {
+        $VAR = @{
+            Uri = $Input.Current
+            OutFile = (Split-Path -Path $Input.Current -Leaf).Split('?')[0]
         }
-        Invoke-WebRequest @params | Out-Null
-        Return $params.OutFile
+        Invoke-WebRequest @VAR
+        Return $VAR.OutFile
     }
 }
 
@@ -26,7 +26,7 @@ Function Install-Program {
             'msi' {
                 & msiexec /passive /package $Input.Current | Out-Host
             }
-            'exe' {
+            Default {
                 & ".\$($Input.Current)" /SP- /VERYSILENT /SUPPRESSMSGBOXES /NORESTART | Out-Host
             }
         }
@@ -35,45 +35,73 @@ Function Install-Program {
 }
 
 Function Build-Project {
+    @(
+        @{
+            Cmd = 'lazbuild'
+            Url = 'https://fossies.org/windows/misc/lazarus-3.6-fpc-3.2.2-win64.exe'
+            Path = "C:\Lazarus"
+        }
+    ) | ForEach-Object {
+        If (! (Test-Path -Path $_.Path)) {
+            $_.Url | Request-File | Install-Program
+            $env:PATH+=";$($_.Path)"
+            ".... install $((Get-Command $_.Cmd).Source)" | Out-Host
+        }
+    }
     $VAR = @{
+        Src = 'src'
         Use = 'use'
-        Cmd = 'lazbuild'
-        Url = 'https://netix.dl.sourceforge.net/project/lazarus/Lazarus%20Windows%2064%20bits/Lazarus%203.6/lazarus-3.6-fpc-3.2.2-win64.exe?viasf=1'
-        Path = "C:\Lazarus"
+        Pkg = 'use\components.txt'
     }
-    Try {
-        Get-Command $VAR.Cmd
-    } Catch {
-        Request-File $VAR.Url | Install-Program
-        $env:PATH+=";$($VAR.Path)"
-        Get-Command $VAR.Cmd
-    }
-    If (Test-Path -Path $($VAR.Use)) {
-        & git submodule update --init --recursive --force --remote | Out-Host
-        $COMPONENTS = "$($VAR.Use)\components.txt"
-        If (Test-Path -Path $COMPONENTS) {
-            Get-Content -Path $COMPONENTS | ForEach-Object {
-                If ((! (& $VAR.Cmd --verbose-pkgsearch $_ | Out-Null)) &&
-                    (! (& $VAR.Cmd --add-package $_ | Out-Null)) &&
+    If (Test-Path -Path $VAR.Use) {
+        If ((Test-Path -Path '.gitmodules') &&
+            (& git submodule update --init --recursive --force --remote)) {
+                ".... [$($LastExitCode)] git submodule update" | Out-Host
+            }
+        If (Test-Path -Path $VAR.Pkg) {
+            Get-Content -Path $VAR.Pkg | ForEach-Object {
+                If ((! (& lazbuild --verbose-pkgsearch $_ )) &&
+                    (! (& lazbuild --add-package $_)) &&
                     (! (Test-Path -Path "$($VAR.Use)\$($_)"))) {
-                        $OutFile = Request-File "https://packages.lazarus-ide.org/$($_).zip"
+                        $OutFile = "https://packages.lazarus-ide.org/$($_).zip" | Request-File
                         Expand-Archive -Path $OutFile -DestinationPath "$($VAR.Use)\$($_)" -Force
                         Remove-Item $OutFile
+                        ".... download package $($_)" | Out-Host
                     }
             }
         }
-        Get-ChildItem -Filter '*.lpk' -Recurse -File –Path 'use' | ForEach-Object {
-            & $VAR.Cmd --add-package-link $_ | Out-Host
+        Get-ChildItem -Filter '*.lpk' -Recurse -File –Path $VAR.Use |
+            ForEach-Object { $Result = @() } {
+                If (& lazbuild --add-package-link $_) {
+                    $Result += ".... [$($LastExitCode)] add dependence $($_)"
+                }
+            } { $Result | Out-Host }
+    }
+    Get-ChildItem -Filter '*.lpi' -Recurse -File –Path $VAR.Src |
+        Where-Object { $_ -notlike 'backup' } |
+        ForEach-Object {
+            $Result = @()
+            $exitCode = 0
+        } {
+            $Tmp = (New-TemporaryFile).Name
+            & lazbuild --build-all --recursive --no-write-project --build-mode='release' $_ |
+                Out-File -FilePath $Tmp -Force
+            If ($LastExitCode -eq 0) {
+                $Result += Get-Content -Path $Tmp | Select-String -Pattern 'Linking'
+            } Else {
+                $exitCode+=1
+                $Result += Get-Content -Path $Tmp | Select-String -Pattern 'Error:', 'Fatal:'
+            }
+            Remove-Item $Tmp
+        } {
+            $Result | Out-Host
+            Exit $exitCode
         }
-    }
-    Get-ChildItem -Filter '*.lpi' -Recurse -File –Path 'src' | ForEach-Object {
-        & $VAR.Cmd --no-write-project --recursive --build-mode=release $_ | Out-Host
-    }
 }
 
 Function Switch-Action {
     $ErrorActionPreference = 'stop'
-    Set-PSDebug -Strict -Trace 1
+    Set-PSDebug -Strict #-Trace 1
     Invoke-ScriptAnalyzer -EnableExit -Path $PSCommandPath
     If ($args.count -gt 0) {
         Switch ($args[0]) {
